chore: Remove native platforms for web-only project (#25)
- removed folders: windows, ios, linux, macos, windows

note: Typically when a new project is made, you can modify the arguments of `flutter create` to also include which platforms you want to include files for. When we made this project, the default is to include all of them. Since now we know we're only going with web, we can safely delete the rest